Add LRU cache and POI oriented optimizations
1. refs #6645 Add a caching mechanism for transliterated names
** Transliterating names is now cached.
** Bunch of small optimizations.
** The C++ functions being called by the POI generic script are now
   the minority of the time taken.

+++ hoot-core/src/main/cpp/hoot/core/io/OgrWriter.cpp
@@ -569,7 +569,7 @@ void OgrWriter::writePartial(const boost::shared_ptr<const hoot::Way>& newWay)
    * Make sure this way has any hope of working (i.e., are there enough spots in the cache
    * for all its nodes?
    */
-  if ( newWay->getNodeCount() > _currElementCacheCapacity )
+  if ((long)newWay->getNodeCount() > _currElementCacheCapacity )
   {
     LOG_FATAL("Cannot do partial write of Way ID " << newWay->getId() <<
       " as it contains " << newWay->getNodeCount() << " nodes but our cache can only hold " <<
